llm_agent/requirements.txt
langchain
langchain-community
sentence-transformers
faiss-cpu
chromadb
tqdm
python-dotenv
pandas
jupyter
streamlit